backend/pyproject.toml
[project]
name = "encoding2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.1",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-ollama>=1.0.1",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pillow>=12.0.0",
    "uvicorn>=0.40.0",
]
